Add Respoke-SDK header to requests
This patch adds a "Respoke-SDK" header to the request, which includes
the version of the Respoke-Java library, the operating system the
request is coming from, and the version of Java the request is coming
from.

The library version is pulled from the package manifest, which is
located in the .jar file at `META-INF/MANIFEST.MF`. This file is
populated by the gradle build, and the version is determined by the
command `git describe --tags` with the 'v' that prepends the tag
stripped off. This means that to get a "clean" version, i.e. one
without the number of commits past the tag and the hash appended
to it, you have to perform the build directly after having tagged
a release in git.

Unfortunately, the manifest information is not available when the
tests are run, so under test the first portion of the Respoke-SDK
header that is sent will render as "null/null". I was unable to
find a workaround for this. It only is an issue at test time... I
was able to test it by pulling the built artifact into a separate
project and performing a request using the library and was able
to verify that the header rendered correctly.

This patch is related to MER-4340.